refactor: move Utils class to util package
