+++ src/main/java/dinkplugin/SettingsValidator.java
@@ -2,6 +2,7 @@ package dinkplugin;
 
 import dinkplugin.notifiers.CollectionNotifier;
 import dinkplugin.notifiers.CombatTaskNotifier;
+import dinkplugin.util.Utils;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import net.runelite.api.Client;
