+++ src/test/java/dinkplugin/IgnoredWorldTest.java
@@ -1,5 +1,6 @@
 package dinkplugin;
 
+import dinkplugin.util.Utils;
 import org.junit.jupiter.api.extension.ExtensionContext;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
