+++ src/test/java/dinkplugin/notifiers/KillCountNotifierTest.java
@@ -1,6 +1,6 @@
 package dinkplugin.notifiers;
 
-import dinkplugin.Utils;
+import dinkplugin.util.Utils;
 import dinkplugin.message.NotificationBody;
 import dinkplugin.message.NotificationType;
 import dinkplugin.notifiers.data.BossNotificationData;
